% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeature.R
\name{BasicQC}
\alias{BasicQC}
\title{Basic QC filtering}
\usage{
BasicQC(input, numSD = 3)
}
\arguments{
\item{input}{A list with entries All, Exons, Meta, and QC as produced by getAll}

\item{numSD}{Number of standard deviations to use for filtering}
}
\value{
The input list after filtering based on QC
}
\description{
Does basic QC filtering on data
}
