% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeature.R
\name{cleanUp}
\alias{cleanUp}
\title{Removes cells}
\usage{
cleanUp(input, to_keep = c(), to_rem = c())
}
\arguments{
\item{input}{A list with entries All, Exons, Meta, and QC as produced by getAll}

\item{to_keep}{A list of samples to keep}

\item{to_tem}{A list of camples to remove}
}
\value{
The input list with bad samples removed
}
\description{
Given the input produced by getAll filters out cells
}
