% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine.juncs.new.R
\name{combinedJunctions}
\alias{combinedJunctions}
\title{Merge junctions}
\usage{
combinedJunctions(dat, minDist = 50)
}
\arguments{
\item{dat}{Data frame of junction count information}

\item{minDist}{The distance to collapse}
}
\value{
A new junction dataframe with collapsed junctions
}
\description{
Merges Junctions together is they are within short distance of one another
}
