% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunIsoTest.R
\name{GeneLevel}
\alias{GeneLevel}
\title{Gene level stats}
\usage{
GeneLevel(res)
}
\arguments{
\item{res}{The list of results from RunISO. Alternately, a data frame with just one such result.}
}
\value{
A list with one entry for each entry in res, with gene level results
}
\description{
Takes results of RunISO and aggregates to gene level using FWER correction.
}
