% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeature.R
\name{getAll}
\alias{getAll}
\title{Load Pipeline Output}
\usage{
getAll(lst, QCFilter = T)
}
\arguments{
\item{lst}{A named list of directories produced by the pipeline (one directory per sample)}

\item{QCFilter}{True if you want to perform QC filtering, False otherwise}
}
\value{
A list of data loaded, with entires All for feature count with introns+exons, Exons for counts with only exons, and QC for QC information
}
\description{
Loads the output of the pipeline into R
}
