% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine.juncs.new.R
\name{loadData_Junc}
\alias{loadData_Junc}
\title{Load Junctions}
\usage{
loadData_Junc(lst)
}
\arguments{
\item{lst}{A list of annotated junction files}
}
\value{
A data frame with junction count information
}
\description{
Loads the annotated junctions produced by the pipeline
}
