% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeature.R
\name{loadFeature}
\alias{loadFeature}
\title{Load Feature Counts}
\usage{
loadFeature(lst)
}
\arguments{
\item{lst}{A name array of count files produced by featureCount}
}
\value{

}
\description{
A function to load a list of count data from featureCount
}
