% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeature.R
\name{loadQC}
\alias{loadQC}
\title{Load PICARD QC data}
\usage{
loadQC(lst)
}
\arguments{
\item{lst}{A name array of files produced by RNASeqMetrics}
}
\value{

}
\description{
A function to load QC from PICARD
}
