% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRSEM.R
\name{loadRSEM}
\alias{loadRSEM}
\title{Load RSEM output}
\usage{
loadRSEM(files, gene = T)
}
\arguments{
\item{files}{A name array of isoform or gene level files from RSEM}

\item{gene}{If true load like genes, else load like isoforms}
}
\value{

}
\description{
A function to load a list of data from RSEM in tximport format
}
