% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSalmon.R
\name{loadSalmon}
\alias{loadSalmon}
\title{Load Salmon output}
\usage{
loadSalmon(files)
}
\arguments{
\item{files}{A name array of sf files from Salmon}
}
\value{

}
\description{
A function to load a list of data from Salmon in tximport format
}
