% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSalmon.R
\name{loadSalmonForDrim}
\alias{loadSalmonForDrim}
\title{Load Salmon in testing format}
\usage{
loadSalmonForDrim(files, gene2trans = NULL, nams = c())
}
\arguments{
\item{files}{A name array of sf files from Salmont}

\item{gene2trans}{A dataframe with one entry (the first) being transcript name, the second being the gene name. If NULL will try to load assuming data from our pipeline.}

\item{nams}{An array of names, if not given just uses files as names. Used as sample names.}
}
\value{
Ready from DRIM or other testing
}
\description{
A function to load a list of data from Salmon as count matrix ready for DTU testing
}
