% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRSEM.R
\name{prepRSEMforDrim}
\alias{prepRSEMforDrim}
\title{Make RSEM DRIM Ready}
\usage{
prepRSEMforDrim(counts, gene2trans)
}
\arguments{
\item{counts}{A matrix of count data with rownames being transcript names}

\item{gene2trans}{A dataframe with one entry (the first) being transcript name, the second being the gene name}
}
\value{
Ready from DRIM or other testing
}
\description{
Makes RSEM data ready for DTU testing
}
