% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSalmon.R
\name{prepSalmonforDrim}
\alias{prepSalmonforDrim}
\title{Make Salmon DRIM Ready}
\usage{
prepSalmonforDrim(counts, gene2trans)
}
\arguments{
\item{counts}{A matrix of count data with rownames being transcript names}

\item{gene2trans}{A dataframe with one entry (the first) being transcript name, the second being the gene name}
}
\value{
Ready from DRIM or other testing
}
\description{
Makes Salmon data ready for DTU testing
}
