% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDEX.R
\name{RunDexIso}
\alias{RunDexIso}
\title{Run DEX-Seq on Isoforms}
\usage{
RunDexIso(
  cts,
  meta,
  form = ~sample + exon + condition:exon,
  form_red = ~sample + exon,
  min_samps_feature_expr = 3
)
}
\arguments{
\item{cts}{A dataframe of counts as produced by the load DRIM functions for RSEM and Salmon. First two columns should be feature_id and gene_id, the transcript and gene names.}

\item{meta}{A dataframe with meta data.}

\item{form}{The formula to use for testing, using columns in meta.}

\item{form_red}{the reduced form that is tested against}

\item{min_samps_feature_expr}{The number of samples that a given trascript must be expressed in.}
}
\value{
A list with 2 entries: the first the DEX results at Isoform leve, the second the results at Gene level
}
\description{
A method to run DEX-Seq for ISOform level DTU testing. Implements the approach (including copying some code) from Love et al 2018 (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6178912/)
}
