% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiffUTR.R
\name{RunDiffUTR}
\alias{RunDiffUTR}
\title{Run DiffUTR}
\usage{
RunDiffUTR(lst = c(), meta = c(), gtf = "", useDEX = F, usediffSpliceDGE = F)
}
\arguments{
\item{lst}{A list of bams}

\item{meta}{A dataframe of meta data}

\item{gtf}{The GTF to use}

\item{useDEX}{If true use DEX}

\item{usediffSpliceDGE}{If true use diffSpliceDGE}
}
\value{
The results of diffUTR
}
\description{
Command to run diffUTR
}
