% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunIsoTest.R
\name{RunISO}
\alias{RunISO}
\title{Run alternate DTU test}
\usage{
RunISO(cts, meta, form, contrasts = c(), numSamp = 4)
}
\arguments{
\item{cts}{A dataframe of counts as produced by the load DRIM functions for RSEM and Salmon. First two columns should be feature_id and gene_id, the transcript and gene names.}

\item{meta}{A dataframe with meta data used for testing. Rownames should be the sample names.}

\item{form}{The formula to use for testing, using columns in meta.}

\item{contrasts}{List of contrasts to test. If not supplied will test all coeficients except the intercept.}

\item{numSamp}{The number of samples a transcript needs to be expressed in to be tested}
}
\value{
Returns a list with the DTU results for each contrast tested.
}
\description{
Runs an alternate test for DTU
}
