% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestIntrons_nb_offset_glmgampoi.R
\name{TestIntrons}
\alias{TestIntrons}
\title{Test for increase/decrease in intron levels}
\usage{
TestIntrons(
  dat_all,
  dat_exon,
  meta,
  form = ~condition,
  test = "conditionKO",
  offset = T
)
}
\arguments{
\item{dat_all}{A gene by sample matrix of counts, including both introns and exons}

\item{meta}{The meta data to use}

\item{form}{The formula to be tested (~condition by default)}

\item{test}{The coefficient to be tested}

\item{offset}{A boolea variable that determines if there should be an offset based on the percent intronic reads in each sample}

\item{dat_ex}{A gene by sample matrix of counts, including onlyexons}
}
\value{
The results of testing each gene for changes in the ratio between the intronic and exonic reads
}
\description{
Tests each gene to see if that gene has more/less introns than expected.
}
