% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTestJunc.R
\name{TestPeakOutput}
\alias{TestPeakOutput}
\title{Test Junction usage changes}
\usage{
TestPeakOutput(
  dat_peak,
  meta,
  form = ~condition,
  dat_gene = NULL,
  numReads = 10,
  toTest = "conditionKO"
)
}
\arguments{
\item{dat_peak}{Peak data from regtools}

\item{meta}{Meta data matching peak data, rownames must match that col names of dat_peak}

\item{form}{Formula to fit model for}

\item{dat_gene}{A gene by sample matrix, if not given produced by collapsing dat_peak by gene}

\item{numReads}{The min number of reads before filtering}

\item{toTest}{The coefficient to test}
}
\value{
A per peak p-value and effect size estimate
}
\description{
Tests for changes between conditions of the use of junctions in each gene.
}
